requirements.txt
geonode==2.4rc2
pygdal==1.10.1
gunicorn==19.3.0
celery
django-celery
